+++ app/src/main/kotlin/ac/mdiq/podcini/storage/database/Feeds.kt
@@ -4,7 +4,7 @@ import ac.mdiq.podcini.PodciniApp.Companion.getAppContext
 import ac.mdiq.podcini.net.sync.SynchronizationSettings.isProviderConnected
 import ac.mdiq.podcini.net.sync.model.EpisodeAction
 import ac.mdiq.podcini.net.sync.queue.SynchronizationQueueSink
-import ac.mdiq.podcini.storage.model.FeedType
+import ac.mdiq.podcini.storage.specs.FeedType
 import ac.mdiq.podcini.shared.getEntityId
 import ac.mdiq.podcini.storage.model.ARCHIVED_VOLUME_ID
 import ac.mdiq.podcini.storage.model.DownloadResult
@@ -18,6 +18,9 @@ import ac.mdiq.podcini.storage.specs.EpisodeState
 import ac.mdiq.podcini.storage.specs.MediaType
 import ac.mdiq.podcini.storage.specs.Rating
 import ac.mdiq.podcini.shared.nowInMillis
+import ac.mdiq.podcini.sources.SourceGatewayClient
+import ac.mdiq.podcini.storage.model.ShareLog
+import ac.mdiq.podcini.storage.model.toEpisode
 import ac.mdiq.podcini.utils.EventFlow
 import ac.mdiq.podcini.utils.FlowEvent
 import ac.mdiq.podcini.utils.Logd
@@ -237,6 +240,28 @@ fun createSynthetic(feedId: Long, name: String, video: Boolean = false): Feed {
     feed.keepUpdated = false
     feed.queue = null
     return feed
+}
+
+suspend fun addClientEpisode(client: SourceGatewayClient, url: String,  toFeed: Feed, log: ShareLog? = null) {
+    val episode = client.withProvider { it.buildEpisode(url)?.toEpisode() }
+    if (episode != null) {
+        val episodes = toFeed.episodes
+        val status = if (episodes.firstOrNull { it.identifyingValue == episode.identifyingValue } != null) ShareLog.Status.EXISTING.ordinal
+        else {
+            episode.id = getEntityId()
+            episode.feedId = toFeed.id
+            upsertBlk(episode) {}
+            EventFlow.postStickyEvent(FlowEvent.FeedUpdatingEvent(false))
+            ShareLog.Status.SUCCESS.ordinal
+        }
+        if (log != null) upsert(log) {
+            it.title = episode.title
+            it.status = status
+        }
+    } else {
+        Loge(TAG, "Error adding media at url: $url")
+        if (log != null) upsert(log) { it.details = "Can not build episode" }
+    }
 }
 
 fun addRemoteToMiscSyndicate(episode: Episode) {