+++ app/src/main/kotlin/ac/mdiq/podcini/storage/model/Feed.kt
@@ -750,17 +750,6 @@ class Feed : RealmObject {
     }
 }
 
-enum class FeedType {
-    RSS,
-    ATOM,
-    YouTube,
-    PeerTube,
-    SoundCloud;
-    companion object {
-        fun fromName(name: String?): FeedType? = FeedType.entries.firstOrNull { it.name == name }
-    }
-}
-
 @Serializable
 data class FeedDTO(
     val id: Long,